+++ task_master.go
@@ -24,6 +24,7 @@ import (
 	"github.com/influxdata/kapacitor/services/httpd"
 	"github.com/influxdata/kapacitor/services/httppost"
 	k8s "github.com/influxdata/kapacitor/services/k8s/client"
+	"github.com/influxdata/kapacitor/services/kafka"
 	"github.com/influxdata/kapacitor/services/mqtt"
 	"github.com/influxdata/kapacitor/services/opsgenie"
 	"github.com/influxdata/kapacitor/services/opsgenie2"
@@ -165,6 +166,9 @@ type TaskMaster struct {
 		Global() bool
 		StateChangesOnly() bool
 		Handler(hipchat.HandlerConfig, ...keyvalue.T) alert.Handler
+	}
+	KafkaService interface {
+		Handler(kafka.HandlerConfig, ...keyvalue.T) (alert.Handler, error)
 	}
 	AlertaService interface {
 		DefaultHandlerConfig() alerta.HandlerConfig